add custom date to ContractChart
