+++ src/components/TransactionsChartPanel.vue
@@ -10,9 +10,10 @@
     </template>
 
     <div class="transactions-chart-panel__container">
-      <Line
-        :options="chartOptions"
-        :data="chartData"/>
+      <line-chart
+        v-if="transactionsStatistics"
+        :statistics="transactionsStatistics"
+        :selected-interval="selectedInterval"/>
     </div>
 
     <chart-controls
@@ -22,88 +23,14 @@
 </template>
 
 <script setup>
-import { Line } from 'vue-chartjs'
-import {
-  CategoryScale,
-  Chart as ChartJS,
-  Legend,
-  LinearScale,
-  LineElement,
-  PointElement,
-  Title,
-  Tooltip,
-} from 'chart.js'
+
 import { storeToRefs } from 'pinia'
-import { DateTime } from 'luxon'
 import { useTransactionsStore } from '@/stores/transactions'
+import LineChart from '@/components/LineChart'
 
 const transactionsStore = useTransactionsStore()
-const {
-  transactionsStatistics,
-} = storeToRefs(transactionsStore)
+const { transactionsStatistics } = storeToRefs(transactionsStore)
 const { fetchTransactionsStatistics } = transactionsStore
-
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  Title,
-  Tooltip,
-  Legend,
-)
-
-ChartJS.defaults.font.family = 'Roboto Mono'
-
-const chartData = computed(() => {
-  return {
-    labels: labels.value,
-    datasets: [{
-      data: stats.value,
-      label: null,
-      cubicInterpolationMode: 'monotone',
-      tension: 0.4,
-      borderColor: '#f5274e',
-      backgroundColor: '#f5274e',
-      pointRadius: 3,
-      pointHitRadius: 20,
-    }],
-  }
-})
-
-const chartOptions = {
-  responsive: true,
-  maintainAspectRatio: false,
-  plugins: {
-    legend: {
-      display: false,
-    },
-    tooltip: {
-      tooltip: {
-        position: 'top',
-      },
-      callbacks: {
-        title: function(context) {
-          return context.label
-        },
-      },
-    },
-  },
-  scales: {
-    y: {
-      border: {
-        display: false,
-      },
-    },
-    x: {
-      grid: {
-        color: function() {
-          return 'transparent'
-        },
-      },
-    },
-  },
-}
 
 const selectedInterval = ref('')
 
@@ -112,28 +39,6 @@ await useAsyncData(async() => {
   return true
 })
 
-const stats = computed(() => {
-  return transactionsStatistics.value?.map(stat => {
-    return stat.count
-  })
-})
-
-const labels = computed(() => {
-  return transactionsStatistics.value?.map(stat => {
-    return formatLabel(stat.startDate)
-  })
-})
-
-function formatLabel(label) {
-  const date = DateTime.fromISO(label)
-
-  if (selectedInterval.value === 'month') {
-    return date.toFormat('yyyy-MM')
-  }
-
-  return date.toFormat('MM-dd')
-}
-
 async function loadTransactionsStatistics({ interval, limit, range }) {
   selectedInterval.value = interval
   const params = range
@@ -141,9 +46,6 @@ async function loadTransactionsStatistics({ interval, limit, range }) {
     : `?interval_by=${interval}&limit=${limit}`
   await fetchTransactionsStatistics(params)
 }
-
-// todo mainnet url
-
 </script>
 
 <style scoped>
@@ -157,5 +59,4 @@ async function loadTransactionsStatistics({ interval, limit, range }) {
     margin-top: var(--space-4);
   }
 }
-
 </style>
